+++ cuda/src/processing/filters/low_pass.cu
@@ -45,7 +45,10 @@ extern "C"{
         int blocksize = configs.blocksize;
 
         if ( blocksize == 0 ){
-            int blocksize_aux  = compute_GPU_blocksize(sizez, configs.total_required_mem_per_slice_bytes, true, BYTES_TO_GB * getTotalDeviceMemory());
+            int blocksize_aux  = compute_GPU_blocksize( sizez, 
+                                                        configs.total_required_mem_per_slice_bytes, 
+                                                        true, 
+                                                        BYTES_TO_GB * getTotalDeviceMemory());
             blocksize          = min(sizez, blocksize_aux);
         }